Consolidate date code to one directory
Instead of having most functions in dates.py, an exception class in
errors.py, and a helper class in util_support/date_disambiguator.py,
this puts everything under one directory.

No functional changes, just moving things around.
