+++ docs/api/developer/augur.dates.errors.rst
@@ -1,0 +1,7 @@
+augur.dates.errors
+==================
+
+.. automodule:: augur.dates.errors
+   :members:
+   :undoc-members:
+   :show-inheritance:
