+++ docs/api/developer/augur.dates.rst
@@ -5,3 +5,8 @@ augur.dates
    :members:
    :undoc-members:
    :show-inheritance:
+
+.. toctree::
+
+   augur.dates.date_disambiguator
+   augur.dates.errors
