+++ docs/api/developer/augur.util_support.rst
@@ -10,7 +10,6 @@ augur.util\_support
 
    augur.util_support.color_parser
    augur.util_support.color_parser_line
-   augur.util_support.date_disambiguator
    augur.util_support.node_data
    augur.util_support.node_data_file
    augur.util_support.node_data_reader
